// src/app/models/sjcs/TurnosItem.ts
export class TurnosItem {
     nombre: String;
	  abreviatura: String;
	  idarea: String;
	  idmateria: String;
	  idzona: String;
	  idzubzona: String;
	  area: String;
	  materia: String;
	  idturno: String;
	  jurisdicciones: String;
	  jurisdiccion: String;
	  zona: String;
	  subzona: String;
	  idpartidapresupuestaria: String;
	  grupofacturacion: String;
	  fechabaja: Date;
	  historico: boolean;
	  nletrados: String;
	  idtipoturno: String;
	  fechamodificacion;
	  usumodificacion: String;
	  validarjustificaciones: String;
	  validarinscripciones: String;
	  designadirecta: String;
	  requisitos: String;
	  idpersona_ultimo: String;
	  descripcion: String;
	  activarretriccionacredit: String;
	  letradoasistencias: String;
	  letradoactuaciones: String;
	  codigoext: String;
	  fechasolicitud_ultimo: String;
	  idsubzona: String;
	  idpersonaUltimo: String;
	  fechasolicitudUltimo: String;
	  visiblemovil: String;
	  idgrupofacturacion: String;
	  idjurisdiccion: String;
	  idguardias: String;
	  idordenacioncolas: String;
	  orden: String;
	  nombrepartidosjudiciales: String;
	  idinstitucion: String;
	  fechavalidacion: String;
	  fechasolicitud: String;
	  nifcif: String;
	  idpersona: String;
	  nombrepersona: String;
	  apellido1: String;
	  apellido2: String;
	  alfabeticoapellidos: String;
	  numerocolegiado: String;
	  fechanacimiento: Date;
	  antiguedadcola: Date;
	  saltos: String;
	  compensaciones: String;
	  fechabajapersona: String;
	  fechaActual;
	  idcomboguardias: String;
	  nombreguardia: String;
	  fechabajaguardia: String;
       constructor() { }
}